Move functions from thread_data_fwd.hpp to thread_pool_helpers.hpp
