+++ libs/threadmanager/src/threadmanager.cpp
@@ -1235,23 +1235,4 @@ namespace hpx { namespace threads {
             }
         }
     }
-
-    ///////////////////////////////////////////////////////////////////////////
-    std::int64_t get_thread_count(thread_state_enum state)
-    {
-        return get_thread_manager().get_thread_count(state);
-    }
-
-    std::int64_t get_thread_count(
-        thread_priority priority, thread_state_enum state)
-    {
-        return get_thread_manager().get_thread_count(state, priority);
-    }
-
-    ///////////////////////////////////////////////////////////////////////////
-    bool enumerate_threads(util::function_nonser<bool(thread_id_type)> const& f,
-        thread_state_enum state)
-    {
-        return get_thread_manager().enumerate_threads(f, state);
-    }
 }}    // namespace hpx::threads
