+++ src/runtime/threads/executors/this_thread_executors.cpp
@@ -22,9 +22,11 @@
 #include <hpx/threading_base/set_thread_state.hpp>
 #include <hpx/threading_base/thread_num_tss.hpp>
 #include <hpx/runtime/threads/executors/manage_thread_executor.hpp>
+#include <hpx/runtime/threads/thread_helpers.hpp>
 #include <hpx/affinity/affinity_data.hpp>
 #include <hpx/runtime/threads/resource_manager.hpp>
 #include <hpx/coroutines/thread_enums.hpp>
+#include <hpx/functional/deferred_call.hpp>
 #include <hpx/timing/steady_clock.hpp>
 #include <hpx/threading_base/thread_description.hpp>
 #include <hpx/basic_execution/this_thread.hpp>
